Se agrego AlumnoServiceImpl

+++ becas/src/main/java/mx/uam/cua/tysi/integracion/becas/repository/AlumnoRepository.java
@@ -3,7 +3,6 @@ package mx.uam.cua.tysi.integracion.becas.repository;
 import mx.uam.cua.tysi.integracion.becas.entity.Alumno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-
-public interface AlumnoRepository extends JpaRepository<Alumno, Integer> {
+public interface AlumnoRepository extends JpaRepository<Alumno, Long> {
 
 }
